Split request file repository
