+++ pkg/repository/request/create_test.go
@@ -1,0 +1,164 @@
+package request
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/DATA-DOG/go-sqlmock"
+	"github.com/google/jsonapi"
+)
+
+func TestCreate(t *testing.T) {
+	db, mock, err := sqlmock.New()
+	if err != nil {
+		t.Fatalf("Unexpected error when opening a stub db connection, error: %s\n", err)
+	}
+
+	cases := []struct {
+		name                string
+		mock                func()
+		req                 jsonapi.Linkable
+		expectedID          int64
+		expectedStatus      string
+		expectedDetails     string
+		expectedBitrate     int64
+		expectedResolutionX int
+		expectedResolutionY int
+		expectedRatioX      int
+		expectedRatioY      int
+		expectedVideoName   string
+		errorPresent        bool
+	}{
+		{
+			name: "Should add request to db",
+			mock: func() {
+				mock.ExpectQuery(fmt.Sprintf("INSERT INTO %s", TableName)).
+					WithArgs(64000, 800, 600, 4, 3, 1, "new_video").
+					WillReturnRows(sqlmock.NewRows([]string{"id"}).
+						AddRow(1))
+
+				mock.ExpectQuery(fmt.Sprintf("SELECT id, status, details, bitrate, resolution_x, resolution_y, ratio_x, ratio_y, video_name FROM %s", TableName)).
+					WithArgs(1).
+					WillReturnRows(sqlmock.NewRows([]string{"id", "status", "details", "bitrate", "resolution_x", "resolution_y", "ratio_x", "ratio_y", "video_name"}).
+						AddRow(1, "original_in_review", "", 64000, 800, 600, 4, 3, "new_video"))
+			},
+			req: &Resource{
+				UserID:      1,
+				Bitrate:     64000,
+				ResolutionX: 800,
+				ResolutionY: 600,
+				RatioX:      4,
+				RatioY:      3,
+				VideoName:   "new_video",
+			},
+			expectedID:          1,
+			expectedStatus:      "original_in_review",
+			expectedDetails:     "",
+			expectedBitrate:     64000,
+			expectedResolutionX: 800,
+			expectedResolutionY: 600,
+			expectedRatioX:      4,
+			expectedRatioY:      3,
+			expectedVideoName:   "new_video",
+			errorPresent:        false,
+		},
+		{
+			name: "Should not add request to db",
+			mock: func() {
+				mock.ExpectQuery(fmt.Sprintf("INSERT INTO %s", TableName)).
+					WithArgs(64000, 800, 600, 4, 3, 1, "new_video").
+					WillReturnRows(sqlmock.NewRows([]string{"id"}))
+			},
+			req: &Resource{
+				UserID:      1,
+				Bitrate:     64000,
+				ResolutionX: 800,
+				ResolutionY: 600,
+				RatioX:      4,
+				RatioY:      3,
+				VideoName:   "new_video",
+			},
+			errorPresent: true,
+		},
+		{
+			name: "With invalid json.Linkable",
+			mock: func() {
+			},
+			req:          &invalidResource{},
+			errorPresent: true,
+		},
+	}
+
+	for _, testCase := range cases {
+		t.Run(testCase.name, func(t *testing.T) {
+			testCase := testCase
+			testCase.mock()
+			repo := NewRepository(db)
+			linkable, err := repo.Create(context.Background(), testCase.req)
+			if err != nil && !testCase.errorPresent {
+				t.Errorf("Unexpected error: %s\n", err.Error())
+			}
+
+			if err == nil && testCase.errorPresent {
+				t.Errorf("Should be error\n")
+			}
+
+			if err == nil {
+				request, ok := linkable.(*Resource)
+				if !ok {
+					t.Fatalf("Invalid type assetrion *reqest.Resource\n")
+				}
+
+				if request.ID != testCase.expectedID {
+					t.Errorf("Invalid id, expected: %d, got: %d\n",
+						testCase.expectedID, request.ID)
+				}
+
+				if request.Status != testCase.expectedStatus {
+					t.Errorf("Invalid status, expected: %s, got: %s\n",
+						testCase.expectedStatus, request.Status)
+				}
+
+				if request.Details != testCase.expectedDetails {
+					t.Errorf("Invalid details, expected: %s, got: %s\n",
+						testCase.expectedDetails, request.Details)
+				}
+
+				if request.Bitrate != testCase.expectedBitrate {
+					t.Errorf("Invalid bitrate, expected: %d, got: %d\n",
+						testCase.expectedBitrate, request.Bitrate)
+				}
+
+				if request.ResolutionX != testCase.expectedResolutionX {
+					t.Errorf("Invalid resolution, expected: %d, got: %d\n",
+						testCase.expectedResolutionX, request.ResolutionX)
+				}
+
+				if request.ResolutionY != testCase.expectedResolutionY {
+					t.Errorf("Invalid resolution, expected: %d, got: %d\n",
+						testCase.expectedResolutionY, request.ResolutionY)
+				}
+
+				if request.RatioX != testCase.expectedRatioX {
+					t.Errorf("Invalid ratio, expected: %d, got: %d\n",
+						testCase.expectedRatioX, request.RatioX)
+				}
+
+				if request.RatioY != testCase.expectedRatioY {
+					t.Errorf("Invalid ratio, expected: %d, got: %d\n",
+						testCase.expectedRatioY, request.RatioY)
+				}
+
+				if request.VideoName != testCase.expectedVideoName {
+					t.Errorf("Invalid name, expected: %s, got: %s\n",
+						testCase.expectedVideoName, request.VideoName)
+				}
+			}
+
+			if err := mock.ExpectationsWereMet(); err != nil {
+				t.Errorf("there were unfulfilled expectations: %s\n", err)
+			}
+		})
+	}
+}
